+++ release-notes.md
@@ -2,7 +2,8 @@
 
 ## 0.4 (WIP)
 
-- Enforce tooltip is a string (fix rendering issue in some weird cases)
+- Added support for [neo4j-spatial plugin](https://github.com/neo4j-contrib/spatial) SimplePoint layers
+- Bug fix: enforce tooltip is a string (fix rendering issue in some weird cases)
 - Fix CI
 - Code refactoring
 
@@ -21,7 +22,7 @@
     - Inform the user when query returns no result
     - Warning when switching from advanced to simple query
     - Warning before deleting a layer (definitive action)
-- Possibility to switch from simple to advanced layer with prefilled query
+- Possibility to switch from simple to advanced layer with pre-filled query
 - Possibility to show the generated query for simple mode layer
 - Display the selected marker color in the layer header (left side bar)
 - Maximum number of points shown on the map is customizable (marker layer)